Fix Issue with Featured Image
For some reason, when creating the Featured Image custom field, every time I set the unique key to featured-image it would switch to feature_image after I saved it.  I am not sure, but I assuming that there have been some changes to the anchorcms code that may produce this issue.  I was able to fix this on my build (0.9.2) by changing the featured-image to featured_image in 2 spots.